--- modules/gltg-bmp/.babelrc
@@ -1,12 +1,0 @@
-{
-  "presets": [
-    "@babel/env",
-    "@babel/flow",
-    "@babel/react"
-  ],
-  "plugins": [
-    "@babel/plugin-proposal-class-properties",
-    "@babel/plugin-proposal-export-default-from",
-    "@babel/plugin-proposal-export-namespace-from"
-  ]
-}
