--- modules/gltg-bmp/CHANGELOG.md
@@ -1,30 +1,0 @@
-# Change Log
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](http://keepachangelog.com/) 
-and this project adheres to [Semantic Versioning](http://semver.org/).
-
-## Unreleased
-- Added new visualization and updated the existing one to use vega
-  [GLGVO-724](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-724)
-- Add top ten practices by area treated to BMP results
-  [GEOD-1364](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1364)
-- Export results as PDF
-  [GEOD-1366](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1366)
-  [GH-42](https://github.com/geostreams/geodashboard/issues/42)
-
-## [3.6.0] - 2020-12-14
-
-### Changed
-- Added support for fetching config from the api
-  [GLGVO-709](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-709)
-- Integrate the results view with `gltg-bmp` api and add a few visualizations
-  [GLGVO-711](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-711)
-
-
-## [0.1.0] - 2020-10-01
-### Added
-- Implemented the basic layout
-  [GLGVO-688](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-688)
-- Added webpack production config
-  [GLGVO-692](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-692)
