--- modules/gltg-bmp/README.md
@@ -1,7 +1,0 @@
-## Great Lakes to Gulf Virtual Observatory - Best Management Practices
-
-### Environment variables
-
-| Variable | Type | Default Dev | Default Prod | Description |
-|----------|------|-------------|--------------|-------------|
-| GEOSERVER_URL | string | https://gltg-dev.ncsa.illinois.edu/geoserver | /geoserver | |
