--- modules/gltg-bmp/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@geostreams/gltg-bmp",
-  "version": "0.1.0",
-  "main": "src/index.jsx",
-  "scripts": {
-    "start": "webpack serve --open --config webpack.dev.js",
-    "build": "webpack --config webpack.prod.js"
-  },
-  "author": "NCSA",
-  "license": "ISC",
-  "dependencies": {
-    "@geostreams/core": "^3.6.2",
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "~4.0.0-alpha.57",
-    "@material-ui/styles": "^4.11.2",
-    "favicons": "^6.2.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.3",
-    "react-router-dom": "^5.2.0",
-    "react-vega": "^7.4.2",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
-    "vega": "^5.19.1",
-    "vega-lite": "^4.17.0"
-  }
-}
