--- modules/gltg-bmp/src/index.jsx
@@ -1,7 +1,0 @@
-// @flow
-import reducers from '@geostreams/core/src/reducers';
-import render from '@geostreams/core/src/render';
-
-import routes from './routes';
-
-render(reducers, routes);
