--- modules/gltg-bmp/src/tests/Tests.jsx
@@ -1,11 +1,0 @@
-// @flow
-import React from 'react';
-import BaseTests from '@geostreams/core/src/tests/BaseTests';
-
-const tests = [];
-
-const Tests = () => (
-    <BaseTests components={tests} />
-);
-
-export default Tests;
