+++ package.json
@@ -10,14 +10,12 @@
     "lint:fix": "eslint -c .eslintrc --ext .js --ext .jsx --ignore-path .eslintignore --fix ./",
     "flow": "flow",
     "build": "webpack --config webpack.prod.js",
-    "clean": "find . -name 'yarn.lock' -prune -exec rm '{}' + && find . -name 'node_modules' -type d -prune -exec rm -rf '{}' +",
-    "bmp:start": "yarn workspace @geostreams/gltg-bmp start",
-    "bmp:build": "yarn workspace @geostreams/gltg-bmp build"
+    "clean": "find . -name 'yarn.lock' -prune -exec rm '{}' + && find . -name 'node_modules' -type d -prune -exec rm -rf '{}' +"
   },
   "author": "NCSA",
   "license": "ISC",
   "dependencies": {
-    "@geostreams/core": "^3.6.0",
+    "@geostreams/core": "^3.6.2",
     "@geostreams/core__old": "^3.6.0",
     "@geostreams/geostreaming": "^3.6.0",
     "@geostreams/gltg__old": "^3.6.0",
@@ -33,8 +31,11 @@
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.3",
     "react-router-dom": "^5.2.0",
+    "react-vega": "^7.4.2",
     "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0"
+    "redux-thunk": "^2.3.0",
+    "vega": "^5.19.1",
+    "vega-lite": "^4.17.0"
   },
   "devDependencies": {
     "@babel/core": "^7.13.15",
@@ -94,8 +95,7 @@
       "modules/*"
     ],
     "nohoist": [
-      "**/@geostreams/gltg__old/**",
-      "**/@geostreams/gltg-bmp/**"
+      "**/@geostreams/gltg__old/**"
     ]
   }
 }
