+++ src/containers/BMP/Results/index.jsx
@@ -16,7 +16,7 @@ import logger from '@geostreams/core/src/utils/logger';
 
 import type { Action as PageAction } from '@geostreams/core/src/actions/page';
 
-import { BMP_API_URL } from '../../config';
+import { BMP_API_URL } from '../config';
 import { BMPContext } from '../Context';
 import { RESULTS, createRequestParams } from './config';
 import Pdf from './Pdf';
